[FE] REFACTOR: 네이밍 ProfileDetail로 수정
