+++ src/frontend/src/components/common/ProfileDetail/index.css.ts
